requirements.txt
pytesseract>=0.3.10
streamlit>=1.28.0
pillow>=9.0.0
opencv-python>=4.6.0
numpy>=1.21.0
python-docx>=0.8.11
